validate frontend

+++ frontend/src/views/login/profile.vue
@@ -51,7 +51,7 @@
               </div>
             </div>
           </div>
-          <div v-if="editactive">
+          <div v-if="editactive" class="mb-4">
             <div class="flex items-center mb-2 gap-2">
               <div class="h-5 w-5 rounded-full bg-orange-500"></div>
               <label class="block text-gray-700 text-lg font-bold">
@@ -59,15 +59,20 @@
               </label>
             </div>
             <input
-              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 mb-4 leading-tight focus:outline-none focus:shadow-outline"
+              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              :class="{'border-red-500': $v.firstname1.$error}"
               id="firstname"
               name="firstname"
               type="text"
               placeholder="Address"
-              v-model="firstname1"
+              v-model="$v.firstname1.$model"
             />
+            <div v-if="$v.firstname1.$error" class="text-red-500 mt-1">
+              <p v-if="!$v.firstname1.required">This field is required</p>
+              <p v-if="!$v.firstname1.maxLength">Firstname must not more than 150 letters</p>
+            </div>
           </div>
-          <div v-if="editactive">
+          <div v-if="editactive" class="mb-4">
             <div class="flex items-center mb-2 gap-2">
               <div class="h-5 w-5 rounded-full bg-orange-500"></div>
               <label class="block text-gray-700 text-lg font-bold">
@@ -75,14 +80,18 @@
               </label>
             </div>
             <input
-              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 mb-4 leading-tight focus:outline-none focus:shadow-outline"
+              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              :class="{'border-red-500': $v.lastname1.$error}"
               id="lastname"
               name="lastname"
               type="text"
               placeholder="Address"
-              v-model="lastname1"
-              
+              v-model="$v.lastname1.$model"
             />
+            <div v-if="$v.lastname1.$error" class="text-red-500 mt-1">
+              <p v-if="!$v.lastname1.required">This field is required</p>
+              <p v-if="!$v.lastname1.maxLength">Lastname must not more than 150 letters</p>
+            </div>
           </div>
           <!-- <div>
             <label class="block text-gray-700 text-lg font-bold mb-2">
@@ -110,7 +119,7 @@
                 v-model="idcard"
                 />
             </div> -->
-          <div v-if="editactive">
+          <div v-if="editactive" class="mb-4">
             <div class="flex items-center mb-2 gap-2">
               <div class="h-5 w-5 rounded-full bg-orange-500"></div>
               <label class="block text-gray-700 text-lg font-bold">
@@ -118,16 +127,20 @@
               </label>
             </div>
             <input
-              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 mb-4 leading-tight focus:outline-none focus:shadow-outline"
+              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              :class="{'border-red-500': $v.tel1.$error}"
               id="tel"
               name="tel"
               type="tel"
               placeholder="Tel."
-              v-model="tel1"
-              
+              v-model="$v.tel1.$model"
             />
+            <div v-if="$v.tel1.$error" class="text-red-500 mt-1">
+              <p v-if="!$v.tel1.required">This field is required</p>
+              <p v-if="!$v.tel1.mobile">The mobile number format is incorrect</p>
+            </div>
           </div>
-          <div v-if="editactive">
+          <div v-if="editactive" class="mb-4">
             <div class="flex items-center mb-2 gap-2">
               <div class="h-5 w-5 rounded-full bg-orange-500"></div>
               <label class="block text-gray-700 text-lg font-bold">
@@ -135,15 +148,20 @@
               </label>
             </div>
             <input
-              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 mb-4 leading-tight focus:outline-none focus:shadow-outline"
+              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              :class="{'border-red-500': $v.address1.$error}"
               id="address"
               name="address"
               type="text"
               placeholder="Address"
-              v-model="address1"
+              v-model="$v.address1.$model"
             />
+            <div v-if="$v.address1.$error" class="text-red-500 mt-1">
+              <p v-if="!$v.address1.required">This field is required</p>
+              <p v-if="!$v.address1.maxLength">Address must not more than 150 letters</p>
+            </div>
           </div>
-          <div v-if="editactive">
+          <div v-if="editactive" class="mb-4">
             <div class="flex items-center mb-2 gap-2">
               <div class="h-5 w-5 rounded-full bg-orange-500"></div>
               <label class="block text-gray-700 text-lg font-bold">
@@ -151,16 +169,24 @@
               </label>
             </div>
             <input
-              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 mb-4 leading-tight focus:outline-none focus:shadow-outline"
+              class="shadow appearance-none border rounded-full w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              :class="{'border-red-500': $v.email1.$error}"
               id="email"
               name="email"
               type="text"
               placeholder="Email"
-              v-model="email1"
+              v-model="$v.email1.$model"
             />
+            <div v-if="$v.email1.$error" class="text-red-500 mt-1">
+              <p v-if="!$v.email1.required">This field is required</p>
+              <p v-if="!$v.email1.email">This email format is incorrect</p>
+            </div>
           </div>
-          <div v-if="!editactive" class="flex justify-center py-5 cursor-pointer" @click="editactive = !editactive; edit()">
-            <img width="60px" src="https://cdn.discordapp.com/attachments/958256273592307722/971016464389771294/edit.png" alt="">
+          <div v-if="!editactive" class="flex justify-center py-5 gap-3">
+            <router-link to="/">
+              <img  width="50px" src="https://cdn.discordapp.com/attachments/958256273592307722/971439206360576090/home.png" alt="">
+            </router-link>
+            <img class="cursor-pointer" @click="editactive = !editactive; edit()" width="50px" src="https://cdn.discordapp.com/attachments/958256273592307722/971443526430752788/newww_edit.png" alt="">
           </div>
           <div class="flex items-center justify-between mt-10 pb-5" v-if="editactive">
             <button
@@ -185,7 +211,14 @@
 <script>
 import axios from "axios";
 import navbar from "../component/navbar.vue";
-
+import { required, email, maxLength} from 'vuelidate/lib/validators'
+function mobile (value) {
+  if(!value){
+    return true
+  }else{
+    return !!value.match(/0[0-9]{9}/)
+  }
+}
 // @ is an alias to /src
 export default {
   name: "Profile",
@@ -205,7 +238,29 @@ export default {
       address1: "",
     };
   },
-  components: {
+  validations:{
+    firstname1:{
+      required,
+      maxLength: maxLength(150),
+    },
+    lastname1:{
+      maxLength: maxLength(150),
+      required
+    },
+    tel1:{
+      required: required ,
+      mobile: mobile
+    },
+    address1:{
+      required,
+      maxLength: maxLength(150)
+    },
+    email1: {
+      required: required,
+      email: email
+    },
+  },
+  components1: {
     Navbar: navbar
   },
   mounted() {
@@ -224,34 +279,37 @@ export default {
       this.$router.push("/");
     },
     commit() {
-      if (confirm('Are you confirm?')){
-        axios
-        .post(
-          `http://localhost:3000/update/account/${this.loginuser.user_id}`,
-          {
-            firstname: this.firstname1,
-            lastname: this.lastname1,
-            tel: this.tel1,
-            email: this.email1,
-            address: this.address1
-          }
-        )
-        .then(response => {
-          console.log(response.data)
-          if (response.data != "error") {
-            localStorage.setItem("user", JSON.stringify(response.data));
-            this.$router.push("/");
-          }
-        })
-        .catch(error => {
-          console.log("error");
-          this.error = error.response.data.message;
-        });
+      if(this.$v.$invalid){
+        this.$v.$touch()
+      }else{
+        if (confirm('Are you confirm?')){
+          axios
+          .post(
+            `http://localhost:3000/update/account/${this.loginuser.user_id}`,
+            {
+              firstname: this.firstname1,
+              lastname: this.lastname1,
+              tel: this.tel1,
+              email: this.email1,
+              address: this.address1
+            }
+          )
+          .then(response => {
+            console.log(response.data)
+            if (response.data != "error") {
+              localStorage.setItem("user", JSON.stringify(response.data));
+              this.$router.push("/");
+            }
+          })
+          .catch(error => {
+            console.log("error");
+            this.error = error.response.data.message;
+          });
+        }
       }
       
     },
     edit(){
-      console.log(1);
       this.firstname1 = this.firstname;
       this.lastname1 = this.lastname;
       this.tel1 = this.tel;
